backend/pyproject.toml
[tool.poetry]
name = "quizgen"
version = "0.1.0"
description = ""
authors = ["Yash Sawlani <yash.sawlani@ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
langchain = "^0.3.13"
ibm-watsonx-ai = "^1.1.26"
python-multipart = "^0.0.20"
transformers = "^4.47.1"
tensorflow = "^2.18.0"
huggingface-hub = "^0.27.0"
tiktoken = "^0.8.0"
pypdf = "^5.1.0"
langchain-ibm = "^0.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
